elisp backend for generating emacs language modes
This elisp backend provides a major mode corresponding to the rules in
an LBNF grammar file. It supports:

* syntax highlighting
* code completion

There are many desirable language mode features that it does not
currently support including intelligent indentation, interaction with an
interpreter, language documentation integration, and third party tool
support.

The backend is invoked with `--elisp`. It generates a .el file
containing elisp code defining the types, functions and keywords of the
language described in a LBNF format. The heuristics for inferring these
three keywords are imprecise, leaving plenty of room for improvement.
For the right hand side of each grammar rule, given by `rhsRule`, they
are:

* If the length of the list of terminal/non-terminal is one, and this
  element is a Right value, then this string value is a language type.

* If the list has more than one element, and the head of the list is a
  Right value, then this string value is a language function.

* If the list has more than one element, then all Right values in the
  tail are language keywords.